Fix Ruby style
