+++ _plugins/collapsible_content.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # Author: jcalcaben@magento.com
 #
 # This custom plugin adds a block tag that wraps the content in a jquery-ui accordian
@@ -6,7 +8,6 @@
 
 module Jekyll
   class Collapsible < Liquid::Block
-
     def initialize(tag_name, title, tokens)
       super
       @title = title.to_s
@@ -15,28 +16,25 @@ module Jekyll
     def render(context)
       site = context.registers[:site]
 
-      if defined? site.find_converter_instance
-        converter = site.find_converter_instance(Jekyll::Converters::Markdown)
-      else
-        converter = site.getConverterImpl(::Jekyll::Converters::Markdown)
-      end
+      converter = if defined? site.find_converter_instance
+                    site.find_converter_instance(Jekyll::Converters::Markdown)
+                  else
+                    site.getConverterImpl(::Jekyll::Converters::Markdown)
+                  end
 
       content = super.strip
       content = converter.convert(content)
-      "<div class=\"collapsible\"><b class=\"collapsible-title\">"+@title+"</b><div class=\"collapsible-content\">"+content+"</div></div>"
+      '<div class="collapsible"><b class="collapsible-title">' + @title + '</b><div class="collapsible-content">' + content + '</div></div>'
     end
   end
 end
 
 Liquid::Template.register_tag('collapsible', Jekyll::Collapsible)
 
-
-
 ## TODO: Come up with cleaner solution
 
 module Jekyll
   class CollapsibleH2 < Liquid::Block
-
     def initialize(tag_name, title, tokens)
       super
       @title = title.to_s
@@ -45,26 +43,23 @@ module Jekyll
     def render(context)
       site = context.registers[:site]
 
-      if defined? site.find_converter_instance
-        converter = site.find_converter_instance(Jekyll::Converters::Markdown)
-      else
-        converter = site.getConverterImpl(::Jekyll::Converters::Markdown)
-      end
+      converter = if defined? site.find_converter_instance
+                    site.find_converter_instance(Jekyll::Converters::Markdown)
+                  else
+                    site.getConverterImpl(::Jekyll::Converters::Markdown)
+                  end
 
       content = super.strip
       content = converter.convert(content)
-      "<div class=\"collapsible collapsible-h2\"><h2 class=\"collapsible-title\">"+@title+"</h2><div class=\"collapsible-content\">"+content+"</div></div>"
+      '<div class="collapsible collapsible-h2"><h2 class="collapsible-title">' + @title + '</h2><div class="collapsible-content">' + content + '</div></div>'
     end
   end
 end
 
 Liquid::Template.register_tag('collapsibleh2', Jekyll::CollapsibleH2)
 
-
-
 module Jekyll
   class CollapsibleH3 < Liquid::Block
-
     def initialize(tag_name, title, tokens)
       super
       @title = title.to_s
@@ -73,15 +68,15 @@ module Jekyll
     def render(context)
       site = context.registers[:site]
 
-      if defined? site.find_converter_instance
-        converter = site.find_converter_instance(Jekyll::Converters::Markdown)
-      else
-        converter = site.getConverterImpl(::Jekyll::Converters::Markdown)
-      end
+      converter = if defined? site.find_converter_instance
+                    site.find_converter_instance(Jekyll::Converters::Markdown)
+                  else
+                    site.getConverterImpl(::Jekyll::Converters::Markdown)
+                  end
 
       content = super.strip
       content = converter.convert(content)
-      "<div class=\"collapsible collapsible-h3\"><h3 class=\"collapsible-title\">"+@title+"</h3><div class=\"collapsible-content\">"+content+"</div></div>"
+      '<div class="collapsible collapsible-h3"><h3 class="collapsible-title">' + @title + '</h3><div class="collapsible-content">' + content + '</div></div>'
     end
   end
 end
